Ticket Realted Apis

+++ src/main/java/com/grapplermodule1/GrapplerEnhancement/service/UserService.java
@@ -193,12 +193,17 @@ public class UserService implements UserDetailsService {
     public Boolean deleteUser(Long userId) {
         try {
             log.info("Delete User Service Called, User Id {}", userId);
+            String email = SecurityContextHolder.getContext().getAuthentication().getName();
             Optional<Users> user = userRepository.findById(userId);
-
             if (user.isPresent()) {
+                if(user.get().getEmail().equals(email))
+                {
+                    log.error("Delete User throws DataIntegrityViolationCustomException");
+                    throw new DataIntegrityViolationCustomException("You can not delete yourself");
+                }
                 userRepository.deleteById(userId);
-                log.info("Delete User Service Returning True User");
-                return true;
+                 log.info("Delete User Service Returning True User");
+                  return true;
             } else {
                 log.error("Delete User throws UserNotFoundException");
                 throw new UserNotFoundException("User Not Found With ID : " + userId);
